log_pipeline_function/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-keyvault-secrets
azure-identity
azure-storage-blob
